requirements.txt
absl-py==0.9.0
alabaster==0.7.12
astor==0.8.1
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
bleach==1.5.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
cubes==1.1
cycler==0.10.0
Cython==0.29.14
decorator==4.4.1
defusedxml==0.6.0
docutils==0.16
entrypoints==0.3
enum34==1.1.6
expressions==0.2.3
gast==0.2.2
google-auth==1.11.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grako==3.99.9
grpcio==1.26.0
h5py==2.10.0
html5lib==0.9999999
idna==2.8
imageio==2.6.1
imagesize==1.2.0
imgaug==0.3.0
importlib-metadata==1.5.0
ipykernel==5.1.4
ipyparallel==6.2.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
Keras==2.0.8
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
nose==1.3.7
notebook==6.0.3
numpy==1.18.1
oauthlib==3.1.0
opencv-python==4.1.2.30
opencv-python-headless==4.1.2.30
opt-einsum==3.1.0
packaging==20.1
pandas==1.0.0
pandocfilters==1.4.2
parso==0.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
plotly==4.5.0
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.11.2
ptyprocess==0.6.0
pudb==2019.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.5.2
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.0
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
Send2Trash==1.5.0
Shapely==1.7.0
six==1.14.0
sklearn==0.0
snowballstemmer==2.0.0
Sphinx==2.3.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
tensorboard==2.1.0
tensorboardX==2.0
tensorflow==1.4.0
tensorflow-estimator==2.1.0
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
torch==1.4.0
torchvision==0.5.0
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.8
urwid==2.1.0
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==0.16.1
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==2.1.0
